fix error display task and combine task file
