--- Helper/Helper/VolunteerViews/AssistanceTasksView.swift
@@ -1,63 +1,0 @@
-//
-//  AssistanceTasksView.swift
-//  Helper
-//
-//  Created by Annie Huynh on 5.4.2022.
-//
-
-import SwiftUI
-
-struct AssistanceTasksView: View {
-    
-    // user results from core
-    @FetchRequest(entity: User.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \User.userId, ascending: true)]) var results: FetchedResults<User>
-    
-    // passed task information based on category
-    @Binding var assistance: [Task]
-    @Binding var userInfo: [User]
-    @Binding var taskInfo: [Task]
-    @Binding var availableTasks: [Task]
-    @Binding var volunteerName: String
-    
-    // function to get helpseeker information
-    func getHelpseeker(task: Task) -> User {
-        for user in results {
-            if (user.userId == task.helpseeker) {
-                return user
-            }
-        }
-        return User()
-    }
-    
-    var body: some View {
-        GeometryReader{geometry in
-            Color("Background")
-                .edgesIgnoringSafeArea(.all)
-            ScrollView{
-                ZStack{
-                    VStack(spacing: 20){
-                        ForEach(assistance) { task in
-                            let helpseeker  = getHelpseeker(task: task)
-                            TaskCard(currentTask: task, helpseeker: helpseeker, userInfo: userInfo, volunteerName: volunteerName)
-                                .frame(width: geometry.size.width * 0.9, height: geometry.size.height * 0.23,alignment: .top)
-                                .background(.white)
-                                .cornerRadius(10)
-                                .shadow(radius: 5)
-                                .padding(.horizontal)
-                                .padding(.leading,5)
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-
-struct AvailableTasksView_Previews: PreviewProvider {
-    static var previews: some View {
-        AssistanceTasksView(assistance: .constant([]), userInfo: .constant([]), taskInfo: .constant([]), availableTasks: .constant([]),volunteerName: .constant(""))
-        
-    }
-}
-
